+++ packages/react-core/src/components/atoms/Badge/Badge.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { View } from 'react-native';
-import { Text, borderRadius, ThemeProp } from '@tecsinapse/react-core';
-import styled from '@emotion/native';
+import { ThemeProp } from '@tecsinapse/react-core';
+import { BadgeStyle, TextStyle } from './styled';
 
 export interface BadgeProps {
   variant?: string;
@@ -18,15 +17,3 @@ const Badge = (props: BadgeProps): JSX.Element => {
 };
 
 export default Badge;
-
-/// implementar variants
-const BadgeStyle = styled(View)`
-  border-radius: ${borderRadius.pill};
-  background-color: ${({ theme, variant }) => theme[variant].main};
-  width: 70px;
-  align-items: center;
-`;
-const TextStyle = styled(Text)`
-  color: ${({ theme }) => theme.success.contrast};
-  font-weight: bold;
-`;
