+++ packages/react-core/src/components/atoms/Button/styled.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/native';
+import { TouchableHighlight } from 'react-native';
+
+export const StyledButton = styled(TouchableHighlight)`
+  padding: 10px;
+  background-color: ${({ theme }) => theme.primary.medium};
+  border-radius: 4px;
+`;
