+++ packages/react-core/src/components/atoms/Card/Card.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import { View } from 'react-native';
-import { Text, Paper, typography, spacings } from '@tecsinapse/react-core';
-import styled from '@emotion/native';
+import { default as CardStyle, TextInfo, TextTitle } from './styled';
 
-export const Card = (): JSX.Element => {
+export interface CardProps {
+  children: JSX.Element;
+}
+
+const Card = ({ children }: CardProps): JSX.Element => {
   return (
     <CardStyle>
       <View>
@@ -17,24 +20,4 @@ export const Card = (): JSX.Element => {
   );
 };
 
-const CardStyle = styled(Paper)`
-  width: 400px;
-  min-height: 120px;
-  max-width: 90vw;
-  max-height: 90vh;
-  margin-bottom: 1rem;
-  padding-top: ${spacings.centi};
-  padding-left: ${spacings.deca};
-`;
-
-const TextTitle = styled(Text)`
-  ${typography.headings.h1};
-  color: black;
-`;
-
-const TextInfo = styled(Text)`
-  padding-top: ${spacings.mili};
-  padding-left: ${spacings.micro};
-  ${typography.text.base};
-  color: black;
-`;
+export default Card;
