+++ packages/react-core/src/components/atoms/Card/index.ts
@@ -1,0 +1,1 @@
+export { default as Card, CardProps } from './Card';
