+++ packages/react-core/src/components/atoms/Checkbox/index.ts
@@ -1,0 +1,1 @@
+export { default as CheckBox, CheckBoxProps } from './CheckBox';
