+++ packages/react-core/src/components/atoms/Checkbox/styled.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/native';
+import { spacings, Text } from '@tecsinapse/react-core';
+import { View } from 'react-native';
+
+export const ViewStyle = styled(View)`
+  flex-direction: row;
+`;
+
+export const TextStyle = styled(Text)`
+  margin-left: ${spacings.mili};
+`;
