+++ packages/react-core/src/components/atoms/Input/index.ts
@@ -1,0 +1,1 @@
+export { default as Input, InputProps } from './Input';
