+++ packages/react-core/src/components/atoms/Paper/styled.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/native';
+import { View } from 'react-native';
+
+export const PaperStyled = styled(View)`
+  background: white;
+  padding: 30px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.4);
+`;
