+++ packages/react-core/src/components/atoms/RadioButton/index.ts
@@ -1,0 +1,1 @@
+export { default as RadioButton, RadioButtonProps } from './RadioButton';
