+++ packages/react-core/src/components/atoms/Switch/Switch.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Story } from '@storybook/react';
-import { SwitchProps, Switch } from './Switch';
+import { SwitchProps, default as Switch } from './Switch';
 
 export default {
   title: 'Switch',
