+++ packages/react-core/src/components/atoms/Switch/index.ts
@@ -1,0 +1,1 @@
+export { default as Switch, SwitchProps } from './Switch';
