+++ packages/react-core/src/components/atoms/Switch/styled.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/native';
+import { View, Switch as SwitchRN } from 'react-native';
+
+export const ViewStyle = styled(View)`
+  flex-direction: row;
+`;
+
+export const SwitchStyle = styled(SwitchRN)`
+  margin: 0px 10px 0px 10px;
+`;
