+++ packages/react-core/src/components/atoms/Text/index.ts
@@ -1,0 +1,1 @@
+export { default as Text, TextProps } from './Text';
