+++ packages/react-core/src/index.ts
@@ -9,8 +9,8 @@ export {
 export { default as Input, InputProps } from './components/atoms/Input/Input';
 export * from './components/styles/definitions';
 export { Paper } from './components/atoms/Paper/Paper';
-export { Card } from './components/atoms/Card/Card';
+export { default as Card } from './components/atoms/Card/Card';
 export { default as Badge, BadgeProps } from './components/atoms/Badge/Badge';
 export { Switch } from './components/atoms/Switch/Switch';
 export { RadioButton } from './components/atoms/RadioButton/RadioButton';
-export { CheckBox } from './components/atoms/CheckBox/CheckBox';
+export { default as Checkbox } from './components/atoms/Checkbox/CheckBox';
