--- packages/react-native-kit/src/components/atoms/button/Button.tsx
@@ -1,4 +1,0 @@
-import { Button, ButtonProps } from '@tecsinapse/react-core';
-
-export { ButtonProps };
-export default Button;
