--- packages/react-native-kit/src/components/atoms/text/Text.tsx
@@ -1,4 +1,0 @@
-import { Text, TextProps } from '@tecsinapse/react-core';
-
-export { TextProps };
-export default Text;
