+++ packages/react-native-kit/src/index.ts
@@ -1,6 +1,5 @@
-export { default as Button } from './components/atoms/button/Button';
-export * from './components/atoms/button/Button';
-export { default as Text } from './components/atoms/text/Text';
-export * from './components/atoms/text/Text';
-export { ThemeProvider } from '@tecsinapse/react-core';
-export { default as Input } from './components/atoms/input/Input';
+export * from '@tecsinapse/react-core';
+export {
+  default as Input,
+  InputPropsNative,
+} from './components/atoms/Input/Input';
