+++ packages/react-web-kit/src/index.ts
@@ -1,5 +1,1 @@
-export { default as Button } from './components/atoms/button/Button';
-export * from './components/atoms/button/Button';
-export { default as Text } from './components/atoms/text/Text';
-export * from './components/atoms/text/Text';
-export { ThemeProvider } from '@tecsinapse/react-core';
+export * from '@tecsinapse/react-core';
